Implemented pawn evolving #55 by shekharcodess
I have added the pawn evolving code when it reaches the last rank to the ChessPiece.java file of ChessEngine/src/main/java/core/pojo/ChessPiece.java
I have not hampered the code that was already in this file.
Code that I added are, I added promptPieceEvolution() method that prompts a player to select a piece for pawn evolution. Updated the getValidMoves() method which checks if the piece is a pawn and has reached the last rank and if it does I have added  the promptPieceEvolution() method to allow the player to select a piece for evolution. If a new piece type is selected, it updates the pieceType of the pawn. Added the Pieces enum and updated the processMoves() method which checks if each possible move is within the bounds of the chessboard (between ranks 0-7 and files 0-7) and adds them to the validMoves list.